wally support
this is not tested yet though technically it should function as wally requires "src" to be a script in order to properly, to my knowledge at least.

i've only added an init so the entire repo structure remains as desired, alternatively you can move all the files under src and remove the init I've created.